% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculatedPredicted.R
\name{calculateLLNPret}
\alias{calculateLLNPret}
\title{Calculates the predicted lower limit of normal based off of inputted parameters}
\usage{
calculateLLNPret(df, param = "FEV1", ref = "GLI")
}
\arguments{
\item{df}{A dataframe with colnames at the minimum age and height}
}
\description{
Assumes that df is a dataframe that contains at least the age and the height
of the patient
}
