% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculatedPredicted.R
\name{calculateMeanPret}
\alias{calculateMeanPret}
\title{Calculates the predicted mean predicted based on inputed paramters}
\usage{
calculateMeanPret(df, param = "FEV1", ref = "GLI")
}
\arguments{
\item{df}{A dataframe containing colnames age and height at the minimum}
}
\description{
Calculates the predicted mean predicted based on inputed paramters
}
