% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculatedPredicted.R
\name{calculatePctPret}
\alias{calculatePctPret}
\title{Calculates the predicted lower limit of normal based off of inputed parameters}
\usage{
calculatePctPret(df, param = "FEV1", ref = "GLI")
}
\description{
Calculates the predicted lower limit of normal based off of inputed parameters
}
