% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewCategorical.R
\name{compareCategorical}
\alias{compareCategorical}
\title{Produces a number of categorical plots based on the categorical demographic inputted and a certain delination}
\usage{
compareCategorical(df, demParam, spiroParam, delim, type = "box")
}
\description{
Produces a number of categorical plots based on the categorical demographic inputted and a certain delination
}
