% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewNumerical.R
\name{compareNumerical}
\alias{compareNumerical}
\title{Produces numerical plots that are segregated based off of certain delinations of a
parameter}
\usage{
compareNumerical()
}
\description{
Produces numerical plots that are segregated based off of certain delinations of a
parameter
}
