% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatData.R
\name{formatData}
\alias{formatData}
\title{Reads a file in as a dataframe and formats the columns of commonly used demographic parameters}
\usage{
formatData(
  path,
  heightLabel = NULL,
  sexLabel = NULL,
  ageLabel = NULL,
  ethLabel = NULL
)
}
\arguments{
\item{path}{path to the dataframe}

\item{heightLabel}{height label}

\item{sexLabel}{sex label}

\item{ageLabel}{age label}

\item{ethLabel}{ethnicity label}
}
\value{
dataframe containing reformated data
}
\description{
A helper function that
}
