% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NHANES3Data}
\alias{NHANES3Data}
\title{Simulated demographic spirometry data using the third National Health and
Nutrition Examination Survey (NHANES III) equations}
\format{
A CSV file with 50 observation of 8 variables
\describe{
\item{id}{Simulated patient id using 4 digits}
\item{gender}{Gender denoted by 1 for male, 2 for female}
\item{age}{Simulated age of patient}
\item{ethnicity}{Ethnicity of patient, denoted by 1 = Caucasian,
2 = African-American, 3 = Mexican-American}
\item{smoking}{Whether or not patient is a smoker, denoted by 1 for yes and 0 for no}
\item{FEV1}{Simulated FEV1 value using NHANES3 equations multiplied by
a random permutation between 0.2 and 1.2}
\item{FVC}{Simulated FEV1 value using NHANES3 equations multiplied by
a random permutation between 0.2 and 1.2}
}
}
\source{
Hankinson, J. L., Odencrantz, J. R., & Fedan, K. B. (1999).
Spirometric reference values from a sample of the general U.S. population.
American journal of respiratory and critical care medicine, 159(1), 179–187.
https://doi.org/10.1164/ajrccm.159.1.9712108
}
\usage{
NHANES3Data
}
\description{
A dataset containing simulated patient demographic and spirometry data using
Nutrition Examination Survey (NHANES III) equations
}
\examples{
NHANES3Data
}
\keyword{datasets}
