% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runspiroview.R
\name{runspiroview}
\alias{runspiroview}
\title{Launch Shiny App for spiroview}
\usage{
runspiroview()
}
\value{
No return value but open up a Shiny page.
}
\description{
Launch Shiny App for spiroview
}
\references{
Silva, A. (2022) TestingPackage: An Example R Package For BCB410H.
Unpublished. URL https://github.com/anjalisilva/TestingPackage.
}
