% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segregateBy.R
\name{segregateBy}
\alias{segregateBy}
\title{Segregates dataframes by a demographic delimiter}
\usage{
segregateBy(df, demParam, segBy, segIsNumeric = FALSE)
}
\arguments{
\item{df}{The dataframe containing the data you want to clean}

\item{demParam}{The demographic factor in question}

\item{segBy}{The demographic delimiter in question. For categorical variables,
simply some categorical term or a vector of terms is expected.
For numerical variables,a singlestring consisting of a numeric value preceded
by an inequality operator is expected. For example ">16".}

\item{segIsNumeric}{Declaring if segBy is a numeric value. Default is FALSE.}
}
\value{
a list of resulting dataset:
\itemize{
\item dataset meeting segBy condition
\item dataset that does not meet segBy condition
}
}
\description{
A function that takes a demographic parameter and divides the dataframe based
off of the relationship to said parameter. For example, if demParam is sex
and segBy is Male, a data frame with only male samples and a data frame with
only female samples will be created.
}
\details{
Assumptions:
\itemize{
\item the provided parameters and delimiters exist in your dataset.
\item the segBy delimiter exists in your dataset under the inputted demParam
}
}
\examples{
exampleDF <- data.frame(id=c(0000, 1111, 2222),
                        gender=c(1, 2, 1),
                        height=c(1.52, 1.83, 1.96),
                        ethnicity=c(1, 2, 1),
                        FEV1=c(2.581, 2.2768045, 0.4895280 ),
                        FVC=c(2.924, 3.0208665, 0.6688253))

Example 1: Using a categorical delimiter
results <- segregateBy(df=exampleDF,
            demParam = "gender",
            segBy = "1",
            segIsNumeric = FALSE)

Example 2: Using a numeric delimiter
results <- segregateBy(df=exampleDF,
            demParam = "height",
            segBy = ">1.80",
            segIsNumeric = FALSE)

}
\references{
Wickham H (2022). \emph{stringr:
Simple, Consistent Wrappers for Common String Operations}.
R package version 1.5.0, \url{https://CRAN.R-project.org/package=stringr}.
}
