% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segregateBy.R
\name{segregateByCategorical}
\alias{segregateByCategorical}
\title{Segregates spirometry data based off of a certain demographic, produces dataframes}
\usage{
segregateByCategorical()
}
\description{
Segregates spirometry data based off of a certain demographic, produces dataframes
}
