% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summarizeAllByCategory}
\alias{summarizeAllByCategory}
\title{Provides summary statistics for a specific demographic value}
\usage{
summarizeAllByCategory(df, demParam, delim, delimIsNumeric = FALSE, spiroParam)
}
\description{
A function that provides a table of summary statistics for a specific
demographic value for the
}
