% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculatedPredicted.R
\name{calculateLLNPret}
\alias{calculateLLNPret}
\title{Calculates the predicted lower limit of normal based off of inputted parameters}
\usage{
calculateLLNPret(df, param = "FEV1", ref = "GLI")
}
\arguments{
\item{df}{A dataframe containing colnames age and height at the minimum}

\item{param}{The spirometric metric of interest: "FEV1", "FVC", "FEV1FVC",
"PEF", "FEF2575", "FEV6", "FEV1FEV6. Default is "FEV1"}

\item{ref}{The reference equation to be used}
}
\value{
A numeric vector or data.frame containing length(param) columns with
rows equal to the number of rows in df
}
\description{
A function that calculates the predicted lower limit of normal (LLN) for an
inputted spirometric value and a dataframe.
}
\details{
Assumptions:
\itemize{
\item df is a dataframe row that contains at least the age and the height
of the patient
\item gender is denoted as 1 or 2 for male or female respectively
\item ethnicity is denoted as a value between 1 to 5
}
}
