% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculatedPredicted.R
\name{calculatePctPret}
\alias{calculatePctPret}
\title{Calculates the percent predicted based off of inputted parameters
Assumptions:
\itemize{
\item df is a dataframe row that contains at least the age and the height
of the patient
\item gender is denoted as 1 or 2 for male or female respectively
\item ethnicity is denoted as a value between 1 to 5
}}
\usage{
calculatePctPret(df, param = "FEV1", ref = "GLI")
}
\arguments{
\item{df}{A dataframe containing colnames age and height and a spirometric column
matching param at the minimum}

\item{param}{The spirometric metric of interest: "FEV1", "FVC", "FEV1FVC",
"PEF", "FEF2575", "FEV6", "FEV1FEV6".}

\item{ref}{The reference equation to be used}
}
\value{
A numeric vector or data.frame containing length(param) columns with
rows equal to the number of rows in df
}
\description{
Calculates the percent predicted based off of inputted parameters
Assumptions:
\itemize{
\item df is a dataframe row that contains at least the age and the height
of the patient
\item gender is denoted as 1 or 2 for male or female respectively
\item ethnicity is denoted as a value between 1 to 5
}
}
\examples{

# Example 1: Calculate LLN of FEV1 from the GLI equations

# Example 2: Calculate LLN of FEV1 from NHANES3

# Example 3: Calculate LLN from a dataframe of more than one sample

}
