% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewNumerical.R
\name{compareNumerical}
\alias{compareNumerical}
\title{Produces numerical plot that is segregated based off of certain delinations of a
numerical or categorical parameter}
\usage{
compareNumerical(
  df,
  demParam,
  spiroParam,
  delim = NULL,
  demColor = "red",
  secondParm = NULL,
  secondDelim = NULL,
  secondColor = "purple",
  includeBestFit = FALSE
)
}
\description{
Produces numerical plot that is segregated based off of certain delinations of a
numerical or categorical parameter
}
