% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewNumerical.R
\name{compareNumerical}
\alias{compareNumerical}
\title{Produces scatterplot with values meeting certain thresholds marked}
\usage{
compareNumerical(
  df,
  demParam,
  spiroParam,
  delim = NULL,
  delimColor = "red",
  secondParam = NULL,
  secondDelim = NULL,
  secondParamIsNumeric = FALSE,
  secondColor = "purple",
  includeBestFit = FALSE
)
}
\arguments{
\item{df}{A dataframe containing each patient demographic data and at least
one mean spirometric value}

\item{demParam}{Demographic numeric category of interest}

\item{spiroParam}{The spirometric parameter of interest as a string}

\item{delim}{An delineation of demParam, must follow
the convention of "operator numeric" for example "<10".
Default is NULL}

\item{delimColor}{The color to color the points meeting delim on the graph.
Default is "red"}

\item{secondParam}{An additional demographic variable of interest that can be
either a numeric or categorical value.}

\item{secondDelim}{A delineation of secondParam. If it is a numerical variable
it must follow the convention of "operator numeric" for example "<10".
Default is NULL}

\item{secondParamIsNumeric}{States whether secondParam is a numeric or
categorical variable}

\item{secondColor}{The color to color the points meeting secondDelim
on the graph. Default is "purple"}

\item{includeBestFit}{A boolean stating if a the line of best fit should be
included in the plot}
}
\value{
A basic ggplot scatterplot that can be built upon as needed
}
\description{
A function that produces numerical plot that is segregated based off of certain
delineations of a numerical or categorical parameter. Up to two delineations
can be added.
}
\details{
Assumption: delim exists within the domain of demParam and secondDelim
exists within the domain of secondParam
}
