% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatData.R
\name{formatData}
\alias{formatData}
\title{Reads a file in as a dataframe and formats the columns of commonly used demographic parameters}
\usage{
formatData(path)
}
\value{
dataframe containing reformated data
}
\description{
Reads a file in as a dataframe and formats the columns of commonly used demographic parameters
}
