% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runspiroview.R
\name{runspiroview}
\alias{runspiroview}
\title{Launch Shiny App for spiroview}
\usage{
runspiroview()
}
\value{
No return value but open up a Shiny page.
}
\description{
A function that launches the Shiny app for spiroview.
The code has been placed in \code{./inst/shiny-scripts}.
}
\examples{
\dontrun{
sprioview::runspiroview()
}

}
\references{
Silva, A. (2022) TestingPackage: An Example R Package For BCB410H.
Unpublished. URL https://github.com/anjalisilva/TestingPackage.
}
