% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segregateBy.R
\name{segregateBy}
\alias{segregateBy}
\title{Segregates dataframes by a demographic deliminter}
\usage{
segregateBy(df, demParam, segBy, segIsNumeric = FALSE)
}
\arguments{
\item{df}{The dataframe containing the data you want to clean}

\item{demParam}{The demographic factor in question}

\item{segBy}{The demographic delimiter in question. For categorical variables,
simply some categorical term is expected. For numerical variables, a string
consisting of a numeric value preceded by an inequality operator is expected.
For example ">16".}

\item{segIsNumeric}{Declaring if segBy is a numeric value. Default is FALSE.}
}
\value{
a list of resulting dataset:
\itemize{
\item dataset meeting segBy condition
\item dataset that does not meet segBy condition
}
}
\description{
A function that takes a demographic parameter and divides the dataframe based
off of the relationship to said parameter. For example, if demParam is sex
and segBy is Male, a data frame with only male samples and a data frame with
only female samples will be created.
}
\details{
Assumptions:
\itemize{
\item the provided parameters and delimiters exist in your dataset.
\item the segBy deliminter exists in your dataset under the inputted demParam
}
}
