% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summarizeAllByCategory}
\alias{summarizeAllByCategory}
\title{Provides summary statistics of spirometry data for a specific demographic value}
\usage{
summarizeAllByCategory(df, demParam, delim, delimIsNumeric = FALSE, spiroParam)
}
\arguments{
\item{df}{A dataframe containing each patient demographic data and at least
one mean spirometric value}

\item{demParam}{Demographic category of interest}

\item{delim}{A value to segregate the data by. If the delimiter is numeric,
that is, it contains a number, it must follow the convention of
an inequality operator followed by a number, for example "<10"}

\item{delimIsNumeric}{Boolean stating if the delimiter contains a numeric
numeric value}

\item{spiroParam}{Vector containing the spirometric parameter of interest
as a string}
}
\value{
Returns a dataframe that contains summary statistics as columns and
spirometry parameters of interest as rows
}
\description{
A function that provides a table of summary statistics for a specific
demographic value for the spirometry parameters inputted. This function
expects a dataframe
}
\details{
Assumption: the delimiter is a value that exists within the the domain of
demParam.
}
\examples{
# Example dataset
exampleDF <- data.frame(id=c(0000, 1111, 2222),
                        gender=c(1, 2, 1),
                        height=c(1.52, 1.83, 1.96),
                        ethnicity=c(1, 2, 1),
                        FEV1=c(2.581, 2.2768045, 0.4895280 ),
                        FVC=c(2.924, 3.0208665, 0.6688253))

Example 1: Summary statistics for a categorical variable

catResults <- summarizeAllByCategory(df=testDF,
                                     demParam="gender",
                                     delim="1",
                                     delimIsNumeric = FALSE,
                                     spiroParam="FEV1")
catResults

Example 2: Summary statistics for a numeric variable

numResults <- summarizeAllByCategory(df=testDF,
                                     demParam="age",
                                     delim=">70",
                                     delimIsNumeric = TRUE,
                                     spiroParam="FEV1")
numResults

}
\references{
Wickham H, François R, Henry L, Müller K, Vaughan D (2023).
dplyr: A Grammar of Data Manipulation. https://dplyr.tidyverse.org,
https://github.com/tidyverse/dplyr.

Wickham H (2022). stringr: Simple, Consistent Wrappers for
Common String Operations. https://stringr.tidyverse.org,
https://github.com/tidyverse/stringr.
}
