% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summarizeAllByCategory}
\alias{summarizeAllByCategory}
\title{Provides summary statistics of spirometry data for a specific demographic value}
\usage{
summarizeAllByCategory(df, demParam, delim, delimIsNumeric = FALSE, spiroParam)
}
\arguments{
\item{df}{A dataframe containing each patient demographic data and at least
one mean spirometric value}

\item{demParam}{Demographic category of interest}

\item{delim}{A value to segregate the data by. If the delimiter is numeric,
that is, it contains a number, it must follow the convention of
an inequality operator followed by a number, for example "<10"}

\item{delimIsNumeric}{Boolean stating if the delimiter contains a numeric
numeric value}

\item{spiroParam}{Vector containing the spirometric parameter of interest
as a string}
}
\value{
Returns a dataframe that contains summary statistics as columns and
spirometry parameters of interest as rows
}
\description{
A function that provides a table of summary statistics for a specific
demographic value for the spirometry parameters inputted. This function
expects a dataframe
}
\details{
Assumption: the delimiter is a value that exists within the the domain of
demParam. One patient per row.
}
