% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewCategorical.R
\name{viewCategorical}
\alias{viewCategorical}
\title{Produces a number of categorical plots based on the categorical demographic inputted}
\usage{
viewCategorical(df, demParam, spiroParam, type = "box")
}
\description{
Produces a number of categorical plots based on the categorical demographic inputted
}
