% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewCategorical.R
\name{viewCategoricalCounts}
\alias{viewCategoricalCounts}
\title{Produces graphs that show the number of samples per demographic category}
\usage{
viewCategoricalCounts(df, demParam, secondParam = NULL, type = "pie")
}
\arguments{
\item{df}{Data that is being analyzed}

\item{demParam}{A categorical demographic variable}

\item{secondParam}{A second categorical demographic variable}

\item{type}{The type of graph desired, options include: pie, doughnut, bar.
Default is pie}
}
\value{
an output graph
}
\description{
Produces graphs that show the number of samples per demographic category
}
