% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewCategorical.R
\name{viewCategoricalCounts}
\alias{viewCategoricalCounts}
\title{Produces graphs that show the number of samples per demographic category.}
\usage{
viewCategoricalCounts(df, demParam, type = "pie")
}
\arguments{
\item{df}{Data that is being analyzed}

\item{demParam}{A categorical demographic variable}

\item{type}{The type of graph desired, options include: pie, bar.
Default is pie}
}
\value{
an output graph
}
\description{
A function that produces graphs visualizing the number of samples per a
categorical demographic variable.
}
\examples{
plotData <- GLIData
# Example 1: One categorical variable
\dontrun{
plot <- viewCategoricalCounts(df=plotData,
                        demParam="ethnicity",
                        secondParam=NULL,
                        type="pie")
}
# Example 2: Two categorical variables
\dontrun{
plot <- viewCategoricalCounts(df=plotData,
                              demParam="ethnicity",
                              secondParam="gender",
                              type="pie")
}
}
\references{
Wickham H (2016). ggplot2: Elegant Graphics for Data Analysis.
Springer-Verlag New York. ISBN 978-3-319-24277-4,
https://ggplot2.tidyverse.org.
}
