% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewNumerical.R
\name{viewNumerical}
\alias{viewNumerical}
\title{Produces a scatterplot visualizing a numeric variable to spirometric variable}
\usage{
viewNumerical(df, demParam, spiroParam, includeBestFit = FALSE)
}
\arguments{
\item{df}{A dataframe containing each patient demographic data and at least
one mean spirometric value}

\item{demParam}{Demographic category of interest}

\item{spiroParam}{The spirometric parameter of interest as a string}

\item{includeBestFit}{A boolean stating if a the line of best fit should be
included in the plot}
}
\value{
A basic ggplot scatterplot that can be built upon as needed
}
\description{
A function a basic ggplot scatter plot visualizing the relationship between a
numeric variable to a spirometric value of interest.
}
\examples{
# Example 1: Basic numeric plot with viewNumerical
\dontrun{
plotData <- GLIData
plot <- viewNumerical(df=plotdata,
                      demParam="height",
                      spiroParam="FEV1",
                      includeBestFit=FALSE)
}
}
