% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewNumerical.R
\name{viewNumerical}
\alias{viewNumerical}
\title{Produces a scatterplot visualizing a numeric variable to spirometric variable}
\usage{
viewNumerical(df, demParam, spiroParam, includeBestFit = FALSE, title = NULL)
}
\arguments{
\item{df}{A dataframe containing each patient demographic data and at least
one mean spirometric value}

\item{demParam}{Demographic category of interest}

\item{spiroParam}{The spirometric parameter of interest as a string}

\item{includeBestFit}{A boolean stating if a the line of best fit should be
included in the plot}

\item{title}{Title for the plot as a character string. Default is NULL}
}
\value{
A basic ggplot scatterplot that can be built upon as needed
}
\description{
A function a basic ggplot scatter plot visualizing the relationship between a
numeric variable to a spirometric value of interest.
}
\examples{
# Example 1: Basic numeric plot with viewNumerical
\dontrun{
plotData <- GLIData
plot <- viewNumerical(df=plotData,
                      demParam="height",
                      spiroParam="FEV1",
                      includeBestFit=FALSE)
}
}
\references{
Wickham H (2016). ggplot2: Elegant Graphics for Data Analysis.
Springer-Verlag New York. ISBN 978-3-319-24277-4,
https://ggplot2.tidyverse.org.
}
