% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewNumerical.R
\name{viewNumerical}
\alias{viewNumerical}
\title{Produces a number of numerical plots based on the numerical demographic inputted}
\usage{
viewNumerical(df, demParam, spiroParam, export = FALSE)
}
\description{
Produces a number of numerical plots based on the numerical demographic inputted
}
